requirements.txt
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
numpy>=1.20.0
matplotlib>=3.4.0
pillow>=8.0.0
tensorboard>=2.7.0
scipy>=1.7.0
tqdm>=4.62.0
gradio>=3.0.0
pandas>=1.3.0